feat: implement AutoLoadPolicy feature on Enforcer, CachedEnforcer, SynchedEnforcer
