+++ enforcer.go
@@ -18,6 +18,8 @@ import (
 	"errors"
 	"fmt"
 	"strings"
+	"sync/atomic"
+	"time"
 
 	"github.com/Knetic/govaluate"
 	"github.com/casbin/casbin/v3/effect"
@@ -45,6 +47,9 @@ type Enforcer struct {
 	autoSave           bool
 	autoBuildRoleLinks bool
 	autoNotifyWatcher  bool
+
+	stopAutoLoad    chan struct{}
+	autoLoadRunning int32
 }
 
 // NewEnforcer creates an enforcer via file or DB.
@@ -236,6 +241,48 @@ func (e *Enforcer) SetEffector(eft effect.Effector) {
 	e.eft = eft
 }
 
+func (e *Enforcer) IsAudoLoadRunning() bool {
+	return atomic.LoadInt32(&(e.autoLoadRunning)) != 0
+}
+
+// StartAutoLoadPolicy starts a go routine that will every specified duration call LoadPolicy
+func (e *Enforcer) StartAutoLoadPolicy(d time.Duration) {
+	// Don't start another goroutine if there is already one running
+	if e.IsAudoLoadRunning() {
+		return
+	}
+	atomic.StoreInt32(&(e.autoLoadRunning), int32(1))
+	ticker := time.NewTicker(d)
+	go func() {
+		defer func() {
+			ticker.Stop()
+			atomic.StoreInt32(&(e.autoLoadRunning), int32(0))
+		}()
+		n := 1
+		log.LogPrintf("Start automatically load policy")
+		for {
+			select {
+			case <-ticker.C:
+				// error intentionally ignored
+				_ = e.LoadPolicy()
+				// Uncomment this line to see when the policy is loaded.
+				// log.Print("Load policy for time: ", n)
+				n++
+			case <-e.stopAutoLoad:
+				log.LogPrintf("Stop automatically load policy")
+				return
+			}
+		}
+	}()
+}
+
+// StopAutoLoadPolicy causes the go routine to exit.
+func (e *Enforcer) StopAutoLoadPolicy() {
+	if e.IsAudoLoadRunning() {
+		e.stopAutoLoad <- struct{}{}
+	}
+}
+
 // ClearPolicy clears all policy.
 func (e *Enforcer) ClearPolicy() error {
 	// TODO: implement ClearPolicy in adapter and move model.ClearPolicy after adapter.ClearPolicy
@@ -420,7 +467,7 @@ func (e *Enforcer) enforce(matcher string, explains *[][]string, rvals ...interf
 
 	eftStream := e.eft.NewStream(e.model["e"]["e"].Value, cap)
 
-	if  policyLen != 0 {
+	if policyLen != 0 {
 		if len(e.model["r"]["r"].Tokens) != len(rvals) {
 			return false, fmt.Errorf(
 				"invalid request size: expected %d, got %d, rvals: %v",
@@ -510,7 +557,7 @@ func (e *Enforcer) enforce(matcher string, explains *[][]string, rvals ...interf
 			eft = effect.Indeterminate
 		}
 
-		eftStream.PushEffect(eft);
+		eftStream.PushEffect(eft)
 	}
 
 	// log.LogPrint("Rule Results: ", policyEffects)
